Minor fixes (bugs and otherwise) to bug_report.sh
Add two missing tools to the TOOLS variable (jsemtblgen and
jsemcgen.sh).

Fix it so that the -h string will be saved to the bug report file as
well.

Change an assignment of EXIT_CODE=10 to just increment it.

Start out EXIT_CODE at 9 so that the help message is consistent with the
actual errors - 10 is the first error code so now if something occurs
it'll be at least 10.