Add version to vermod.sh
The fact it had no version is indeed ironic but I was not sure what to
set it to. It was arbitrarily selected based on the commits. The initial
release was obviously 0 and there seems that there was one other update
that was code related so I chose 0.1. I backdated the version to the
date of that commit.

The -V flag shows the version and exits (same code as printing help) and
the version is printed in the usage string as well.

The man page was updated to add this option as well as improve the
formatting of the options (use .TP instead of .PP).