Add version to usage string in run_{flex,bison}.sh
Rearranged the -V in the usage string so that it comes right after the
-h as that seems a better place because the rest of the options will not
immediately exit. I've done this for some other scripts as well possibly
without noting it in the git log.

Updated the man pages as well to use better formatting for options
(making sure to also update the order of the -V option).