Minor revisions to bug_report.sh wrt exit codes
Although it is convenient I have changed back parts of commits
96d0c32c30a666573696eebc639898e23406d404 and
324149a5cddf353b7744ae7e0f226dd0637b192b. The exception is that for a
non-executable tool the exit code will be shared. I believe that with
the messages we can say that exit code (as it is right now it's) 19
means that one or more tools cannot be executed. If it turns out that
it'd be better to check each one individually this can be done but
later.