Release bug_report.sh version 0.3 - make modular
In order to make the script much easier to maintain it is now much more
modular. Instead of having to repeat basically the same code over and
over again it is in a single loop (or two if you count the check for
execute bit and running -h on each tool). The loop also controls the
exit code as well as the command number (see below) so that it's easy to
add a new command in the list even if one of the commands has to be
somewhere in the list (one would have to update the base codes i.e. the
index but that wouldn't take much time). To add a command all one has to
do is add an element to the new COMMANDS array (making sure to double
quote it) and add a comment as to what it is and its purpose.

This might or might not be called experimental but I think it's a good
change with perhaps one downside (not sure it is a downside though): you
can't go to a specific line in the script and determine what command
will run.  However you can easily identify which command it is in the
array simply by the index and one could easily update the loop so it
says what command number it is.

Also the script will say what is about to be run and this seems like
a vast improvement as it makes the script much easier to maintain and as
a bonus it's more compact as well. If however there is reason to not
have it this way I have thought of another way to help with the exit
code issue as well. It would not be as modular but it would at least
make it a bit easier to identify the command to run (in fact, as noted
above, this commit adds what amounts to an identification number for
each command which I think is also a bonus).

Finally the script now calls ./run_bison.sh -v 1 and ./run_flex.sh -v 1.

I believe that makes comment
https://github.com/ioccc-src/mkiocccentry/issues/248#issuecomment-1153283608
almost entirely complete. The only things left to be done (besides the
things not explicitly stated because they weren't thought of yet) is
that not all scripts have versions in the help/usage string. These can
be added later.