Add version to jstr_test.sh
Just like the other scripts that did not originally have a version this
was arbitrarily selected based on git log. However in the case of this
one I had to do

    git log --follow jstr_test.sh

as it has had two previous names. Just like the other scripts I have
given the date of the commit that seems like the most recent commit
worthy of a new version which is also arbitrary.

Updated jstr_test.8.