Typo fix in Makefile
I'm almost certain this was a search and replace error on my part some
days back when I renamed the script from bug-report.sh to bug_report.sh.
Anyway the Makefile referred to bug_report.sh.sh in the hostchk_warning
which is wrong and would return an error as no such file exists.