Improve -h messages related to exiting, update iocccsize
Instead of mentioning the specific exit value for -h and -V,
just mention that it will exit.  Let the Exit Codes for -h
and the EXIT STATUS, if needed, mention the actual exit code
in order to reduce the change that the exit code mentioned
will be wrong.

Added -V documentation to jnum_chk.8.

Fixed -h output for iocccsize.

Changed iocccsize version from '28.11 2022-07-10'
to '28.12 2022-10-17'.