Improve comment about struct dyn_array
First of all I made the comment reflect the name of the struct (it said
'array' and now it says 'dyn_array').

It also had said 'a dynamic array of identical elements' but this can be
read as each element in the array is exactly the same thing so I changed
it to say instead 'a dynamic array of elements of the same type' which
is clearer (though perhaps there's still room for improvement).