Fix - to \- in man pages
Because I could not think of a way at this hour with such poor sleep to
do it directly I just removed \ before all '-'s and then added them back
like:

    sed -i -e 's/\\-/-/g' -e 's/-/\\-/g' *.[138]

I then updated the date of the man page. This was started because of a
mkiocccentry commit ce061e0e160004d2de7d8e5fd145bcb79ab7eb48:

    Made -h, -V and other -option in man pages consistent

    We are NOT sure of this is the best way, but
    when a -option is mentioned in the EXIT STATUS section
    of the man page, we put a \ in front of it.

    Example:

    \-h and help string printed or \-V and version string printed

That is indeed the correct way. I fix them as I notice them but this
time I just did a mass replace as above.

In all likelihood this will have to be done again but for now they're
all correct.