Minor fix to commit f7e24a436f
I did a git commit --amend but forgot to specify bug_report.sh so the
updates did not actually go through. Thus the final changes in the git
log are now being committed. Specifically the part about making it
clearer what part of the log file is from what command.