Improve bug_report.sh
The script now has sections as suggested by Landon on GitHub. There is
probably room for improvement in the section titles and/or commands
within each section but it's a good start. This did involve changing
some exit codes and moving commands about the file.

I also tried to make the format a bit more consistent (there's
definitely room for improvement here). For instance each section is
prefixed by '# SECTION N: ...' where '...' is an all upper case
one line description of the section. Then for each command run it's
prefixed by '##' and so on. This is not perfect yet but it will help
with making it easier to check for specific problems.

On the subject of problems if an issue is found it's prefixed with:

    ### ISSUE FOUND:

and there happens to be an Easter egg with this in the script. If you
want a hint you just have to know that I'm a punster and (naturally) a
logophile.

There are other formatting updates as well but once again there's
certainly room for improvement.