Typo fixes in run_bison and run_flex
The backup files referred to in run_bison.sh and run_bison.8 were not
correct. It referred to the bison generated files rather than the backup
files.

The backup file referred to in run_flex.sh and run_flex.8 was incorrect.
It referred to the flex generated file rather than the backup file.
Additionally run_flex.sh only refers to one file so it should not have
said files at all.

Additionally I moved the man pages to section 1 as everyone using this
repo will be using the scripts (most likely via the Makefile). Also when
I eventually make a new repo for the JSON parser (as Landon suggested) I
might end up using these scripts as well. The Makefile has been updated
for this and both the file names and the section in the .TH macro(?) in
the man pages have been updated to be 1 instead of 8.