Typo fix iocccsize_test.sh + clarify shellcheck
There were a few typos in the script that were fixed but I also went
ahead and clarified why we have to disable a specific shellcheck
warning.

This commit is part of a set of commits meant to clean up scripts in the
sense of shellcheck warnings.  In some cases I was able to remove the
need to disable the warning but in the cases where I could not (for
instance if it's not possible) I give a bit more information (where
there was none).