Add shellcheck warning triggered in run_usage.sh
With this commit my round of clean up with shellcheck is finished.
In some of these commits I have fixed the warnings and in the other
cases I have given more information on why the checks are disabled.

However as for the comments go there will be further updates for sure
specifically more consistency (formatting and wording).