Add -S to not prepend apology to bison/flex files
The default is to prepend the apology file but when I make a new repo
for the parser (as Landon suggested) I might want this feature for the
scripts. Of course I might decide I don't as I do actually find the
generated code offensive but the option is there now for anyone who
doesn't want it.

The man pages were updated for this option as well as removing this from
list of bugs.