New version of jparse_test.sh - tests files too
The script has been expanded to test files under test_JSON/good and
test_JSON/bad. The script still checks at least one file that has line
based json blob(s) but now it checks files too.

This should help when doing more rigorous tests on the parser as all one
needs to do is add good and bad files to the correct subdirectory and
the script will test each file found (searches *.json via find(1)).

An important note is that there are some files with a NUL byte which
should be (eventually) changed to use uuencode/uudecode as one of the
GitHub issues talks about but this can be done later.

I've updated the man page and the version (as the summary of this log
suggests) of the script.

There were four bad files and two good files added to test_JSON/bad and
test_JSON/good respectively. Many more could be added.

An important note about this. That directory already has two
subdirectories for info.json and author.json files. These are more for
chkentry but it seemed like these new files should also be under
test_JSON. The difference is that this is for testing the parser itself
rather than a tool that uses the parser to validate that the .info.json
and .author.json files are good or bad. One could certainly move these
new files to a different directory if this is so desired but it seemed
like it would be better to share the directory.