Use -v 7 for run_{bison,flex}.sh in bug_report.sh
As suggested in the GitHub issue page.

I also updated the message letting the user know that they can (if they
so desire) edit the log bug it's discouraged. Specifically I also noted
it has user name. Maybe this is not necessary. If so it can be changed
back but I did these changes yesterday quickly and I need to take a
break now. I don't think the script is alarming: it's just giving them
an example of what they might wish to edit out (or not).