Mass man page update
Added comments at the top of the man pages. There is some humour
(including irony) in all of them but some are more specialised or
specific. One I would like to mention is that dbg.3 is dedicated to
Grace Hopper who popularised the term debugging after a real moth had
got stuck in a mainframe which caused it to malfunction (the term, to my
understanding, already existed, but she made it popular after this
incident).

There's some fun with feathers and tar in the txzchk.1 and txzchk_test.8
man pages of course. The rest are probably more or less the same.