Update bug_report.sh - new checks
Several new tools have been added to the list of checks but the list is
incomplete. This is because under macOS and other BSD Unices many
commands do not have a --version or similar option. There are a variety
of ways we can attempt to get something but this will have to come at a
later date. At this point the other tools can also be added (though I
might end up adding the remaining tools to check even if it also has the
problem of no version option).

I also added to the list of commands to run cpp -dM to get a list of
predefined macros.

I also reorganised the sections a bit. Specifically there's now a C
environment section.