New version of hostchk.sh
Added test source files that must return specific values (and in the
case of one print a certain string).

It's not modular yet. Whether this should be done or not I don't know.
One benefit from that might be that we be that it would be easier to run
all of these tests even if one fails. Currently when one fails no
further tests are run.

There are several kinds of tests. First we obtain the value of ENOENT
and then verify that setting errno = ENOENT will indeed be the value
previously obtained. We then check for the string: "errno: $ENOENT"
(whatever ENOENT might be: keep it in a variable for consistency check
but also in case some system has a different value). We also print
"Hello, world" and check that that was indeed printed. There are checks
for return values 1 and 0 and maybe another one or two I've forgotten.

I'm not sure if this is all that needs to be done but it might be (other
than possibly making it modular).

Updated .gitignore: added hostchk/. Even though I explicitly remove it
when the script finishes and also via trap it's now in the .gitignore
just in case something goes wrong.