Modularise hostchk.sh
Now if a file fails to compile it will not exit but instead move on to
the next test. If a test fails it will not exit either but rather move
on to the next test.

It might be that the script is finished but if it's not it's close to
being finished I think.