Comments consistency update
Many files have had header comments updated to be more consistent. The
files in particular all refer to JSON in some way or another. Some
changes I will specifically refer to:

In some cases I changed 'this' to 'the' in 'This JSON scanner (or
parser) was co-developed by ...'. This is because of the way it flows
wrt the rest of the text. In other cases I kept 'this'.

I changed most instances of 'JSON scanner' to 'JSON parser'. The
exception is the scanner itself. For jparse.h I said 'scanner and
parser' as it's for both. Perhaps they should all say 'scanner and
parser' rather than 'parser' but for now it'll do.

In some cases I added 'the concept of this file was developed by Landon
Curt Noll' and in others I slightly changed the wording to flow a bit
better (it's hoped).

I tried to make the order of the same or similar text the same in all
the files but there probably are a few differences. Nevertheless this is
good progress on consistency when it comes to the json parser and files
that make use of it.