Change cnt -> count
There are two reasons for this. First is so that it spells a word and
does not trigger vim spellcheck (a bonus is it also is more
descriptive). The other is it's too lexicographically close or similar
to another word that one might accidentally type (or read) if one is
tired or distracted. :-) This is probably not good.

Of course it might be wrong for me to write that in the log but I'm not
sure where else to put it so there it is.