Checkpoint fixed test_url
Birthday v6.2 checkpoint :-)

Fixed the test_url() test as it was too cleaver to be clear
that is was wrong.  While the URL test is naive, it is
probably good enough to do what we need to do without trying
to parse a generalized URI.  As of this commit, the function
does enough of a sanity check for our purposes.