Fixes for shellcheck under CentOS
The following issues were resolved for bug_report.sh:

In some cases it's okay to quote the COMMAND variable. In others we have
to disable the warning because shellcheck is totally wrong and quoting
it would break the script.

The following issues were resolved for hostchk.sh:

In what seems like a brain dead problem with shellcheck under CentOS
(fine under macOS) it could not parse instances like:

    cat <<EOF>"$WORK_DIR/pre-errno.c"

and instead needed:

    cat <<EOF >"$WORK_DIR/pre-errno.c"

I'm not sure why this is and maybe there's a good reason but the script
worked even under CentOS so I suspect not. Either way the error was
quite ironic, stating that there was a parser error in a simple command.
It triggered SC1009, SC1073 and SC1072 and trying to disable them did
not work. Adding a space solved it.

The following issues were fixed for iocccsize_test.sh:

In what appears to be another brain dead problem of CentOS shellcheck it
complained that there was no space after a '=' and a comment, stating:

    warning: Remove space after = if trying to assign a value (for empty string, use var='' ... ). [SC1007]

so this was disabled for both lines.