Fix mkiocccentry_test.sh and update chkentry
The zero length rule 2a and rule 2b sizes will no longer return false.
However I did note in a comment (in both functions) that currently the
override booleans are not checked and my guess is this needs to be
added. However as I'm not as clear on the json semantics testing I've
not done this. Nevertheless chkentry now passes with 0 rule_2a_size and
rule_2b_size so mkiocccentry_test.sh now passes. Besides the override
check I believe this makes chkentry entirely complete.

I've updated the comments for chkentry as well.

The man page was updated to no longer say it's not completed.