ioccc_test.sh now writes to a log file
This might or might not be necessary but I added it prior to commit
ea6407583b002ad51401b3c6a1ca797c995ac351 in order to have a log of what
went wrong.

The only thing I'm not sure about is that bug_report.sh will run make
test and pipe it to tee -a with the log file of bug_report.sh and now
ioccc_test.sh will also write text to stdout and the new log file. It
appears to be okay but I'm not sure if this is an actual problem. If it
is I will roll back this commit but if not I think it'll be useful for
when running the script directly (or just make test).

Another option would be to change bug_report.sh to not do this with make
test and instead have it cat that log file to the bug report file but
I'll worry about that if necessary.