Resolve #239 - remove binary data
New version of jparse_test.sh.

Instead of having binary data files in the repo, jparse_test.sh now uses
uudecode(1) on uuencoded files. If uudecode is not found the tests are
skipped. Files which have the extension .uu are expected to be uuencoded
files.

Since there's no need for binary files in the good subdirectory I've
only searched the bad subdirectory.

Updated the man page.

For this to work the function is_open_stream() had to change a bit
namely that if the stream is a tty or (in our case more importantly)
stdin we return true. It's possible to close stdin but without this we
get an illegal seek and so the change to jparse_test.sh will not work.