Undo need for uudecode
Per discussions on GitHub I've rolled back the need for uudecode as the
need for it is no longer valid after all. Thus the *.uu files are
removed and the earlier files are back.

I also undid the man page changes as the -u option no longer is there.

Updated the good and bad directories (for general json files) to be
under test_JSON/general.json to be clearer what it's for: the info.json
is for .info.json files, the author.json is for .author.json file (which
will be, it seems, changed to be .auth.json) and the general.json is for
any json file.