Made some soup
This is an experimental test for using a sub-directory for
the chkentry code that is not part of the top level main code.

We employed some hacks within the Makefile, with only minor hacks
required in 2 header files.  We had to adjust the #include
lines in both the soup/chk_sem_auth.h and soup/chk_sem_info.h files.

We are not intending to fully address #338 at this time.
We only tried one method related to the `chkentry(1)` code
because it was relatively self-contained AND it had a number
of files that were more "soup-like" than others. :-)

There may be better ways to do this, and if so we would
welcome such a better method.  Meanwhile, lets us see how
we like this approach before we consider working on
the larger scope for issue #338.

This approach might not be best for when it comes to the
future _dyn_array_ and the _json parser_ repos.  We can
experiment with the existing _dbg_ repo when it comes time
to ponder now to do that.  However all that is for later, not now.