Fix trap in txzchk_test.sh
There were actually two places where trap was used because the use of
mktemp was used in two different places (the first place used it three
times and the second one time).

I also explicitly added removal of all the temporary files at the end of
the script. This should also be done for any other scripts that use
temporary files if they don't already.