Finish #239
Although the issue was closed and technically resolved yesterday, there
still was the issue where test_JSON/{auth,info}.json/bad had files that
were invalid JSON. It was agreed that this was no longer correct: the
files should be valid JSON but invalid for the .info.json and .auth.json
files.

The file deletions were done via:

    for i in * ; do ../../../jparse "${i}" 2>/dev/null; if [[ "$?" -ne 0 ]]; then git rm "${i}" ; fi ; done|grep 'is bad'

A note of interest is one file at least has a NUL byte but this was not
removed. Why? Because it's in a string which is valid. I'm not sure if
this file will actually trigger an error with chkentry or not as it
would depend on what else is in the file that might be wrong (unless we
do check for NUL bytes in strings - cannot recall - in the check
functions). If it's not checked this might need to be added.