Add chkentry_test.sh script to test chkentry
It is almost the same as part of jparse_test.sh except it acts on the
directories test_JSON/{auth,info}.json/{bad,good}.

Man page added.

Makefile updated to add the script to be checked by shellcheck as well
as to install (and check with checknr) the man page.

An important point is there seem to be some files in the subdirectories
that fail the test. I do recall that at one point earlier on we
discussed that there might be invalid test files and I suspect that
that's what is happening.

I know Landon was working on a test suite for chkentry too so if this is
not what is needed I am happy for a different script to be used.