Bug fix test_manifest() wrt case of filenames
As discussed a while back (and is in fact done in txzchk) when counting
filenames we have to use case-insensitive string comparison. This fixes
the other bad test cases failing (chkentry now rejects it), specifically
if two extra files are EXTRA and extra it now fails as it was agreed to
a while back.