Add all back all json files
Due to a miscommunication I had removed the invalid JSON files from
{info,auth}.json/bad rather than move them to general.json. Because I
need to get going I am adding all the files from {info,auth}.json/bad to
general.json/bad and the duplicates can be removed later on. But this
way the directory has all the files that were removed - as well as some
others.