Updated chkentry_test
We updated chkentry_test.sh to incorporate the requirements for
error reporting and exit codes from our internal test tool.
The majority of the chkentry_test.sh was left in place.
Mostly the error reporting and exit code processing was changed.

Updated chkentry_test(8) man page according to the above.

Sequenced exit codes in chkentry.c (that we forgot in the
previous commit).