Temporarily undo moving of man pages
Unfortunately I forgot about the rules checknr and man2html when I did
the move and as I do not have the time to do a proper fix I'm making it
so at least all the rules will work. I have saved a backup of the
changes so I can do this at a later date.