Fix jparse_test.sh
The problem happened because of a series of incidents as follows.
Originally the idea for the {auth,info}/bad was for both invalid data
because it was invalid for chkentry (or originally jauthchk and
jinfochk) but the other day it was decided that the files should be
valid JSON but fail the semantics tests (only).

A while back I had the idea (to test the general json parser) of having
other files that would not be tested by (what became) chkentry_test.sh
but only jparse_test.sh. Now yesterday when I deleted the invalid JSON
files it was decided they should be moved to the general.json
subdirectory.

However as I had to hurry off in my head I decided to copy all the bad
files to the bad directory since clearly they were bad. That's the
problem. But some of the files are actually valid JSON so they should
actually be in the good subdirectory!

Additionally I made a slight change to jparse_test.sh: add a summary
(for string failures and file failures) so I could easily extract a list
of files. To do this more easily I prefixed each line with either 'FILE:'
or 'STRING:'. However this prefix could probably be removed now. Given
that it helped with fixing the problem I have kept it in for now.