Remove unused function prototypes
The functionality for ignoring warnings has been not needed for quite
some time. The functions were removed but the prototypes were not.

I also updated comments so that any reference to 'being co-developed by'
was changed to 'was co-developed by' (or if both chkentry and the JSON
parser 'were'). I'm aware of some files that still need some of this
added but those that had it in some form should be fixed.