Added gross hack to look for invalid bytes before parsing
On the assumption that JSON does not allow for raw non-encoded NUL
bytes, we added a XXX - gross hack - XXX to scan the data block
for invalid characters.

At issue according to the so-called JSON spec, are bytes between
0x00 and 0x1f EXCEPT for \n (newline), \r (return) and \t (tab).
If any such bytes are detected, then an error is raised BEFORE the
scanner/parser is invoked.  While it is true that \n, \r and \t
must be JSON encoded, it is reasonable that such characters might
be found in a normal data file, so we will not object if we find
those 3 characters not encoded in the data block.

As a quick and dirty hack, we "manually" scan the data block
for such invalid bytes and object whenever we find them.

TODO: We need to better handle this, perhaps in the scanner?