Improved NUL byte pre-scan of data prior to JSON parsing
To avoid parser / scanner problems, data blocks are pre-scanned
for NUL bytes.  If any NUL byte is detected in the data block,
the JSON parser / scanner is NOT invoked.  Instead, error message(s)
are issued and the data block is declared invalid JSON.

We added a static nul_scan() function to the parser code that will
both detect NUL bytes and report on the first MAX_NUL_REPORTED
occurrences of a NUL byte (with line number and byte position within
the line).

We limit NUL byte reporting to the first MAX_NUL_REPORTED occurrences
in case a large block of NUL bytes are encountered, in order to
avoid producing a large flood of NUL byte error messages.