Expanded JSON string regular expression
We expanded the JSON string regular expression (JSON_STRING) from:

    \"([^\n"]|\\\")*\"

to:

    \"([^"\x01-\x1f]|\\\")*\"

We did NOT include \x00 (NUL byte) because of possible problems
with NUL bytes such as at the end of C-style strings.
Because the nul_scan() function is called before the parser code
is invoked, we do NOT need to complicate the scanner with
NUL bytes in regular expressions.