Add check for IOCCC submit server readiness
If the submit server is not ready the user of mkiocccentry is told that
because the submit server is not ready they may not enter the contest.
It also asks them to please not email the Judges. If however the server
is ready it tells them how to register as well as (after forming the
tarball that's not a test entry) where they can upload their entry
(again assuming they've registered).

This commit makes these checks a bit more modular (well - the checks
weren't actually in before except for test mode but the display of these
things were there and it is this that is modular but the check for the
submit server being ready was added to it in the two new functions).