Add check for has_nul in json_decode()
This commit is on a different branch because it's somewhat hackish and I
might be missing something. More specifically I'm not quite sure it's
correct. Although the has_nul is set to true if it's encoded there very
possibly is a better way. Not only that but I might have missed some
cases.

Nevertheless with a file like:

    $ cat nul.json
    "test\u0000test"

    one will see:

    $ ./jparse -J 5 nul.json
    in parse_json_stream(): JSON debug[5]: calling parse_json on data block with length 17:
    in yyparse(): JSON debug[5]: under json_string: returning $json_string type: JTYPE_STRING
    JSON tree[5]:	lvl: 0	type: JTYPE_STRING	len{q0a}: 9	value:	"test\0test"
    in yyparse(): JSON debug[5]: under json_value: returning $json_value type: JTYPE_STRING
    JSON tree[5]:	lvl: 0	type: JTYPE_STRING	len{q0a}: 9	value:	"test\0test"
    in yyparse(): JSON debug[5]: under json_element: returning $json_element type: JTYPE_STRING
    JSON tree[5]:	lvl: 0	type: JTYPE_STRING	len{q0a}: 9	value:	"test\0test"
    in yyparse(): JSON debug[5]: under json: starting: json: json_element
    in yyparse(): JSON debug[5]: under json: $json_element type: JTYPE_STRING
    in yyparse(): JSON debug[5]: under json: about to perform: $json = $json_element;
    in yyparse(): JSON debug[5]: under json: returning $json_element type: JTYPE_STRING
    JSON tree[3]:	lvl: 0	type: JTYPE_STRING	len{q0a}: 9	value:	"test\0test"
    in yyparse(): JSON debug[5]: under json: about to also perform: *tree = $json;
    in yyparse(): JSON debug[5]: under json: ending: json: json_element
    in parse_json(): JSON debug[1]: valid JSON
    valid JSON

This output is correct but whether anything else is missing I do not
know. But that's why it's a separate pull request.

Note that the only time has_nul != NULL is in the function
json_conv_string().

There were also some typo fixes.