Fix have_timegm not executable in bug_report.sh
The problem appears to be that have_timegm is not by default compiled.
It's compiled by have_timegm.sh so bug_report.sh now checks if that can
be executable.

This does make me wonder if it should be modified to also run
have_timegm itself since for some systems that might be important. That
can come at another commit.

Hopefully this resolves the other part of #393.