Fix and improve use of mktemp
In some cases on some platforms, mktemp requires the file
template to have X's in it.

Changed the use of mktemp so that in addition to .XXXXXXXXXX
in the middle, there is also a fixed string suffix on the
temporary file.

Updated .gitignore to ignore files created by mktemp.
Updated make clobber rule to remove such files.