Fix and improve fmt_depend.sh, add comments
The make depend rule's keeping the contents above the
dependency lines has been simplified.  The previous version
of the Makefile, Makefile.orig, is removed if the make
depend process did not change the Makefile.

Added a message to inform the user if the Makefile
was modified and that Makefile.orig is the original.