Improve error messages in jparse
Added locations to error messages! Now when a syntax error is
encountered one might see for example:

    $ ./jparse test/test_JSON/general.json/bad/auth.missing-line.70.json
    syntax error at line 69 on column 3: }
    ERROR[1]: ./jparse: invalid JSON

I chose this file in particular as it does show a problem: if you look
at that file it looks like it's actually at column 17. This is because
tabs are there and are being counted as one. This should possibly be
fixed but maybe not.

The next step would be to add filenames to the error message but this is
an improvement anyway.