Fix run_usage.sh
Add supporting of man directory as man pages are no longer found in the
root directory of this repo.

As a reminder this tool is just a helpful hack but when everything is in
order save man pages there will likely be many updates to the man pages
that will be harder to verify and this tool could become less useful.
It's also worth noting that if the usage line crosses lines it would
only detect the first line so the usage string has to be temporarily
modified to account for this.

Updated the version: not the version number but the date.

Updated the man page as well.