Improve invalid token warning at lexer level
For instance a file that has:

    $ cat test.json
       &test

would show:

    Warning: yylex: at line 1 column 4: invalid token: 0x26 = <&>
    syntax error at line 1 on column 4: &
    ERROR[1]: ./jparse: invalid JSON

Using yyerror() here, while possible, is problematic because of the
ordering of the code in that function. It might also not be as helpful
anyway since at the lexer level we're simply warning even if it's an
invalid token.

It's important to note that unfortunately strlen("\t") == 1 so columns
can be incorrect according to what one might see (and what might be
detected in an editor) and the YYLTYPE structure member first_column. If
there's a solution to this I don't yet know it.