Include filename in json parser errors (yyerror())
We now include the filename in error messages! This is unfortunately (at
least at this time) done via a global variable declared in jparse.h
(along with other globals that we are forced to use). This could be
considered a work in progress, maybe.

There is one thing I'm not so sure about but if necessary I will change
it at another time (must leave after this). Fortunately it's quite easy
to change. This thing is that prior to calling parse_json_file() (which
takes a char const *filename) we set filename to what we used to pass
into the function directly and then use filename instead. That takes
place in two locations: jsemtblgen.c and jparse_main.c. The possible
problem is that parse_json_file() has a parameter 'filename' so this
might or might not be considered a problem even if it's actually the
same variable. In all places it's a char const *filename.

Should this be a problem it can be easily fixed. Since we rely on other
global variables (until such a time - if this ever comes to pass - that
the scanner also becomes re-entrant) the global nature of filename
should not (at least according to my mind at this hour ? :-) ) be a
problem. If it becomes a problem this commit can be either fixed or
rolled back but it is ideal to have the filename.

A word about stdin: we of course use the typical - and that's exactly
what's printed in that case. I thought about whether or not it should
say "stdin" instead but I decided against that because it might suggest
that there is a file called stdin in the current directory. Of course
the same could be said for '-' itself as it's also possible to have a
file by that name. However that is less likely than 'stdin'. Another
option is possibly _stdin_ or something like that but that can be
decided later.

One more note of filename being global. The parser does not allow more
than one filename or string so there should be no problem of files
changing. However since I specifically first set 'filename' to the
correct argument (passed into main) this wouldn't be a problem anyway.
Thus if we ever do decide to allow for more than one string or file this
should not become a problem.

With this commit I believe error messages are complete as far as
functionality goes (this could change as we discover additional issues
of course). I have tested it with several files. One might note that in
some cases the filename is not printed. An example is when there are
invalid bytes in the file. Another example is an empty file. I believe
in the latter case it's because the action of setting the
yylloc->filename is never allowed to happen since nothing was scanned.
This is just a guess on my part and might or might not be the case. As
for the invalid bytes it's because the parser is never called at this
point: it's a fatal error if invalid bytes are found.

Whether error recovery should be added is another matter but I believe
that that is out of scope of this repo (this might also be out of scope
but it's been bugging me for a while that filenames are not shown so I
finally added it).